move createComputePlan to new file
