+++ chaincode/tuple_test.go
@@ -18,8 +18,6 @@ import (
 	"encoding/json"
 	"testing"
 
-	"github.com/stretchr/testify/require"
-
 	"github.com/stretchr/testify/assert"
 )
 
@@ -93,58 +91,6 @@ func (stub *myMockStub) saveWrittenState(t *testing.T) {
 	return
 }
 
-func TestCreateComputePlan(t *testing.T) {
-	scc := new(SubstraChaincode)
-	mockStub := NewMockStubWithRegisterNode("substra", scc)
-	myStub := myMockStub{MockStub: mockStub}
-	myStub.saveWhenWriting = true
-	registerItem(t, *mockStub, "algo")
-	myStub.MockTransactionStart("42")
-	myStub.saveWhenWriting = false
-
-	// Simply test method and return values
-	inCP := defaultComputePlan
-	outCP, err := createComputePlan(NewLedgerDB(&myStub), assetToArgs(inCP))
-	assert.NoError(t, err)
-	assert.NotNil(t, outCP)
-	assert.EqualValues(t, outCP.ComputePlanID, outCP.TraintupleKeys[0])
-
-	// Save all that was written in the mocked ledger
-	myStub.saveWrittenState(t)
-
-	// Check the traintuples
-	traintuples, err := queryTraintuples(NewLedgerDB(&myStub), []string{})
-	assert.NoError(t, err)
-	assert.Len(t, traintuples, 2)
-	require.Contains(t, outCP.TraintupleKeys, traintuples[0].Key)
-	require.Contains(t, outCP.TraintupleKeys, traintuples[1].Key)
-	var first, second outputTraintuple
-	for _, el := range traintuples {
-		switch el.Key {
-		case outCP.TraintupleKeys[0]:
-			first = el
-		case outCP.TraintupleKeys[1]:
-			second = el
-		}
-	}
-	assert.NotZero(t, first)
-	assert.NotZero(t, second)
-	assert.EqualValues(t, first.Key, first.ComputePlanID)
-	assert.EqualValues(t, first.ComputePlanID, second.ComputePlanID)
-	assert.Len(t, second.InModels, 1)
-	assert.EqualValues(t, first.Key, second.InModels[0].TraintupleKey)
-	assert.Equal(t, first.Status, StatusTodo)
-	assert.Equal(t, second.Status, StatusWaiting)
-
-	// Check the testtuples
-	testtuples, err := queryTesttuples(NewLedgerDB(&myStub), []string{})
-	assert.NoError(t, err)
-	require.Len(t, testtuples, 1)
-	testtuple := testtuples[0]
-	require.Contains(t, outCP.TesttupleKeys, testtuple.Key)
-	assert.EqualValues(t, second.Key, testtuple.Model.TraintupleKey)
-	assert.True(t, testtuple.Certified)
-}
 func TestSpecifiqArgSeq(t *testing.T) {
 	t.SkipNow()
 	// This test is a POC and a example of a test base on the output of the log
